job_share/requirements.txt
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cymem==1.31.2
Cython==0.28.4
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
Django==2.0.2
django-appconf==1.0.2
django-bootstrap-form==3.4
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-filter==2.0.0
django-model-utils==3.1.2
en-core-web-md==2.0.0
Faker==0.8.13
idna==2.7
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
numpy==1.15.0
packaging==17.1
pathlib2==2.3.2
pinax-messages==2.0.2
pinax-templates==2.0.1
plac==0.9.6
pluggy==0.7.1
preshed==1.0.1
py==1.5.4
pyparsing==2.2.0
pyreadline==2.1
python-dateutil==2.7.3
pytz==2018.4
regex==2017.4.5
requests==2.19.1
six==1.11.0
spacy==2.0.12
sqlparse==0.2.4
text-unidecode==1.2
thinc==6.10.3
toolz==0.9.0
tqdm==4.24.0
ujson==1.35
urllib3==1.23
virtualenv==16.0.0
websockets==3.2
wincertstore==0.2
wrapt==1.10.11
